refactor: rename destination-pyairbyte-universal to destination-smoke-test
