+++ airbyte-integrations/connectors/destination-smoke-test/destination_smoke_test/run.py
