+++ airbyte-integrations/connectors/destination-smoke-test/main.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2024 Airbyte, Inc., all rights reserved.
 """Main entry point for the destination connector."""
 
-from destination_pyairbyte_universal.run import run
+from destination_smoke_test.run import run
 
 
 if __name__ == "__main__":
